rename folder annotators -> annotation
