+++ moviebot/nlu/annotation/rule_based_annotator.py
@@ -10,10 +10,10 @@ from copy import deepcopy
 from nltk import ngrams
 from nltk.corpus import stopwords
 
-from moviebot.nlu.annotators.slot_annotator import SlotAnnotator
-from moviebot.nlu.annotators.item_constraint import ItemConstraint
-from moviebot.nlu.annotators.operator import Operator
-from moviebot.nlu.annotators.slots import Slots
+from moviebot.nlu.annotation.slot_annotator import SlotAnnotator
+from moviebot.nlu.annotation.item_constraint import ItemConstraint
+from moviebot.nlu.annotation.operator import Operator
+from moviebot.nlu.annotation.slots import Slots
 
 
 class RBAnnotator(SlotAnnotator):
