+++ moviebot/nlu/data_loader.py
@@ -5,7 +5,7 @@ __author__ = "Javeria Habib"
 import json
 import os
 
-from moviebot.nlu.annotators.slots import Slots
+from moviebot.nlu.annotation.slots import Slots
 
 
 class DataLoader:
